Refactored ps6000a drivers to match with psospa and ps4000a common driver calls.
